Allow hooks to replace generic snap assertion
The generic assertion will be put in place if there are preseeded snaps.
A hook that wants to add its own model should be able to overwrite
the generic assertion.  This patch allows a hook to continue unchanged
calling snap_prepare_assertions with a custom brand/model.  The patch
will replace the generic brand/model if a custom brand/model is specified.
Prior to this patch the generic brand/model would remain in place.