Use snap cli rather than custom snap-tool
snap-tool was added to support a deprecate cohort-key feature of the snap store.
Recent changes in snap assertions have added additional fields which snap-tool
is not retrieving.  This resulted in snap install failures on first boot.

This patch removes snap-tool and returns to using the snap cli.  This ensures
snap downloads will function without odd incompatibilities.