999-cpc-fixes: enable more code on grub2 armhf & arm64
armhf & arm64 images use grub. And despite disk-image &
disk-image-uefi installing all the grubs, some of the configuration is
done in the 999-cpc-fixes. Specifically removal of "quiet splash" is
done there, but not active on armhf & arm64. This results in arm
images to boot with "quiet splash".

Enable running the later portions of 999-cpc-fixes on armhf & arm64.

Drop duplicate call to update-grub, as update-grub2 is simply a
symlink to update-grub.

Add a guard around the call to reconfigure grub-pc, to only do that
when it is installed.

This makes armhf & arm64 uefi images consistent with amd64 uefi
images.

LP: #1925780