auto/config: preserve germinate-output directory when removing config
Germinate doesn't take very long at all to run but downloading the
indices it operates on can take a while and nothing else in auto/config
does so not doing it every time you run "lb config" can be a real time
saver.

The code that invokes germinate already checked if the output was
already there but it was unconditionally deleted by the time control got
to that point.