RISC-V: use GRUB on Unmatched board
Up to now we have used u-boot-menu for preinstalled images for the SiFive
HiFive Unmatched and Unleashed boards and GRUB for all other RISC-V images.
The choice was made because RISC-V GRUB was not available when the SiFive
boards where released.

Let the Unmatched and Unleashed board preinstalled images use GRUB.
Simplify the code.

Signed-off-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>