Enhance unminimize to transform into a base image for ubuntu-cpc
The unminimize script previously just restored the system documentation
and translations, man pages, and installed ubuntu-minimal and ubuntu-standard
packages to provide the familiar Ubuntu minimal system. But such an image
never became an equivalent of base image.

Upon investigation and looking at how the base image is constructed -
https://git.launchpad.net/livecd-rootfs/tree/live-build/auto/config#n1108 -
we use the following things:
 - minimal task
 - standard task
 - cloud-image task (which involves ubuntu-server)
 - ubuntu-minimal package
 - server task if arch != amd64

OTOH, in the unminimize script, we use the following:
(https://git.launchpad.net/livecd-rootfs/tree/live-build/auto/build#n286)
 - ubuntu-minimal package
 - ubuntu-standard package

So upon running some tests, it was found that if we install ubuntu-server
(with --fix-policy flag), we get the resulting image equivalent to that
of a base image.

cf: https://warthogs.atlassian.net/browse/CPC-3033