Drop live-build/ubuntu-server/includes.chroot.ubuntu-server-minimal/etc/hosts
/etc/hosts in an installed system comes from one of several places:
- the netbase package which ships a sane default
- scripts/casper-bottom/18hostname, which is mostly identical to the netbase
  one but sets a hostname to the flavor name
- the installer which writes one to the target fs
- chroot includes in livecd-rootfs

A chroot include is only needed for images that don't include netbase (so
don't inherit from the minimal seed), don't boot casper, and don't run an
installer.

NONE of these conditions are true for the ubuntu-server live image.  And in
addition, the /etc/hosts here differs from the others.

So, remove it.