fix: armhf flavour for 6.5+ kernel is now generic (LP: #2033677)
With the introduction of the 6.5 kernel for mantic on 13th September ago we are seeing image build failures
on the armhf builds. The build failure was `No kernel output for generic-lpae!`.

Introduced in the 6.4 kernel and therefore now also in 6.5 there is no generic-lpae flavor anymore. it's just generic now.

As such this commit updates the expected flavour for armhf to generic.