fix: ensure minimize-manual script uses passed in argument for all commands (LP: #2036198)
minimize-manual takes an argument of the path to the chroot but of the 3 commands that are run the argument
is only used twice and with the third hardcoding "chroot" as the path to the chroot.

Thankfully "chroot" has been the patch passed in for the current uses of minimize-manual but this
could cause issues later if that were to change.

This commit resolves the problem preventing future issues.