fix: arm flash-kernel removal from ubuntu-cpc project cloud images (LP: #2037074)
With the migration from linux-kvm to linux-virtual/generic for the mantic minimal
images we can now start building arm64 minimal cloud images.

When building initial test images we noticed that the flash-kernel package was being
installed. This is not required for EFI images.

This commit removes the flash-kernel package from the cloud images arm installs

The only images built using the livecd-rootfs ubuntu-cpc project for arm64 and armhf
which are not cloud images (which therefore do not require flash-kernel) are the
raspi images. raspi does require flash-kernel. But the raspi images use the
`ubuntu-server-raspi` task to install the flash-kernel. As such this non raspi
flash-kernel package install can be completely removed.

It also means that initramfs-tools and dracut-install are installed as these are dependencies
of flash-kernel package.