Setup preinstalled pool in preinstalled-pool.sources
Also Use Signed-By: /etc/apt/keyrings/preinstalled-pool.gpg and
make sure we only update from that .sources file as we did before.

This code may all be dead, who can say.

FIXME: We should figure out how to do an armored export of that key
and then embed it in the signed-by field instead of using a keyring
file.