fix: `return 0` instead of `return` if the intent is not to exit build
If the previous if statement checking if PASSES_TO_LAYERS is true,
then the last return code be non-zero and a return statement with no
argument will return the error code of the if statement thus exiting
the script. This is  not our intent. So we need to return 0 here when
layer name as already been registered