Enable systemd as part of image creation
Microsoft offically support systemd now and our Pro service relies on
it. This option is enabled by default via our launcher (Windows-side) on
first run for quite a while.
Remve this file creation from it, don’t ship it in a package as the
file may be altered by the user to add additional options and ship it as
part of the rootfs.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>