Pare down 'BASE_SEED' handling to only those flavors which still use it.
Per the comments, BASE_SEED was initially used to identify the seed in the
flavor to use for identifying preseeded snaps, and later was also used to
identify which "minimal-remove" seed to apply to an image.

The first usage is now obsolete after a refactor; we now correctly detect
snaps from any of the included seeds without needing an explicit
declaration.

The second usage only applies to installer images that are NOT using layered
squashfs, since for these images 'minimal' is a separate squashfs layer
rather than a list of packages to remove after the fact.

Refactor this code to eliminate pointless definitions of BASE_SEED and
define it only for the subset of flavors today that:
 - have a 'minimal-remove' seed
 - are not using layered squashfs.